pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cuh-resp-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash-compose>=2023.3.12",
    "dash-iconify>=0.1.2,<0.2",
    "dash-mantine-components>=0.15.1,<0.16",
    "humanize>=4.11.0",
    "nbformat>=5.10.4",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "scipy>=1.15.2",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2,<3",
    "ipykernel>=6.29.5",
    "isort>=5.13.2,<6",
    "matplotlib>=3.10.0",
    "pylint>=3.3.3,<4",
]

[tool.autopep8]
max_line_length = 100
in-place = true
recursive = true
aggressive = 1

[tool.isort]
line_length = 100

[tool.pylint.main]

# Files or directories to be skipped. They should be base names, not paths.
ignore = [".venv", "test"]

# Use multiple processes to speed up Pylint. Specifying 0 will auto-detect the
# number of processors available to use, and will cap the count on Windows to
# avoid hangs.
jobs = 8


[tool.pylint.basic]

# Good variable names which should always be accepted, separated by a comma.
good-names = ["i", "j", "k", "ex", "Run", "_"]

# Maximum number of characters on a single line.
max-line-length = 100

[tool.pylint."messages control"]

# Disable the message, report, category or checker with the given id(s).
disable = [
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-locals",
    "bare-except",
    "broad-exception-caught",
    "wildcard-import",
    "unused-wildcard-import"]
